wc - rename component and properties, events and observers
